requirements.txt
google-generativeai
requests


##transformers
##torch
##langchain
##langchain-huggingface